+++ src/framework/Kernel/Plugin.php
@@ -27,7 +27,7 @@ class Plugin {
 	/**
 	 * Plugin version.
 	 */
-	public const VERSION = '1.3.2';
+	public const VERSION = '1.3.3';
 
 	/**
 	 * Singleton instance.